feat. 현재위치-병원위치 계산하여 출력

+++ app/src/main/java/com/zzb/whichhospital/presentation/model/Hospital.kt
@@ -13,5 +13,5 @@ data class HospitalInfo(
     val hospX: Double = 0.0,
     val hospY: Double = 0.0,
     val hospAddr: String = "",
-    val hospDistance: Double = 0.0,
+    val hospDistance: Long = 0L,
 )
